392.03 Fold in stepper motor disable code from G2

+++ firmware/tinyg/config_app.h
@@ -240,7 +240,7 @@ typedef struct cfgParameters {
 	// system group settings
 	float junction_acceleration;	// centripetal acceleration max for cornering
 	float chordal_tolerance;		// arc chordal accuracy setting in mm
-	uint32_t motor_disable_timeout;	// seconds before disabling motors
+	float motor_disable_timeout;	// seconds before disabling motors
 //	float max_spindle_speed;		// in RPM
 
 	// hidden system settings
